+++ packages/airnode-examples/integrations/config-utils.ts
@@ -91,8 +91,8 @@ export const addAirnodeRrpV0DryRunAddress = async (config: Config) => {
 
 export const generateConfigFile = async (dirname: string, config: Config, generateExampleFile: boolean) => {
   const filename = generateExampleFile ? 'config.example.json' : 'config.json';
-  const updatedConfig = await addAirnodeRrpV0DryRunAddress(config);
-  const formattedConfig = format(JSON.stringify(updatedConfig, null, 2), { parser: 'json', printWidth: 120 });
+  const updatedConfig = generateExampleFile ? config : await addAirnodeRrpV0DryRunAddress(config);
+  const formattedConfig = await format(JSON.stringify(updatedConfig, null, 2), { parser: 'json', printWidth: 120 });
   writeFileSync(join(dirname, filename), formattedConfig);
 
   cliPrint.info(`A '${filename}' has been created.`);