+++ packages/airnode-examples/integrations/relay-security-schemes/create-config.ts
@@ -90,7 +90,7 @@ const createConfig = async (generateExampleFile: boolean): Promise<Config> => ({
   templates: [],
   ois: [
     {
-      oisFormat: '2.1.0',
+      oisFormat: '2.3.1',
       title: 'Relay Security Schemes via httpbin',
       version: '1.0.0',
       apiSpecifications: {