+++ packages/airnode-validator/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @api3/airnode-validator
 
+## 0.13.0
+
+### Minor Changes
+
+- [#1888](https://github.com/api3dao/airnode/pull/1888) [`1da62631`](https://github.com/api3dao/airnode/commit/1da62631905cf4b49266f248c8c385b5106d4c4d) Thanks [@dcroote](https://github.com/dcroote)! - Bump OIS to v2.2.0 and make operationParameter optional within endpoint parameters
+
+### Patch Changes
+
+- [#1893](https://github.com/api3dao/airnode/pull/1893) [`cefc5e4a`](https://github.com/api3dao/airnode/commit/cefc5e4abcc0f10a9b0b84b9bf55af2f34ffe1bd) Thanks [@dcroote](https://github.com/dcroote)! - Bump ois to v2.2.1
+
+- Updated dependencies [[`87cee037`](https://github.com/api3dao/airnode/commit/87cee0372afc60acb141ad308d3664172f3cbdb6), [`d2e5a04b`](https://github.com/api3dao/airnode/commit/d2e5a04bf6e88de1888f044dfb171344171ba0ea), [`1d7e4b2f`](https://github.com/api3dao/airnode/commit/1d7e4b2fe4467cee05a6d5f4b34b772d377337df), [`b447fcc5`](https://github.com/api3dao/airnode/commit/b447fcc5d82f63c9393e2ef5651cedf66809a4a3)]:
+  - @api3/airnode-protocol@0.13.0
+
 ## 0.12.0
 
 ### Minor Changes